\name{currentYear}


\title{Current Year}


\alias{currentYear}


\description{

    A variable with the current year.

}


\usage{
currentYear
}


\value{

    returns the current year, a numeric variable.

}


\note{

    It is not allowed to change this variable.
    
}


\examples{
## currentYear -
   currentYear
}


\keyword{chron}

