\name{show-methods}
\alias{show-methods}

\docType{methods}

\alias{show,timeDate-method}
\alias{show,timeInterval-method}


\title{Show methods}

\description{
  Show methods for \code{"timeDate"} objects. 
}

\section{Methods}{
  \describe{
    \item{\code{signature(object = "timeDate")}}{
	Print function for objects of class \code{"timeDate"}.
    }
    \item{\code{signature(object = "timeInterval")}}{
	%%  ~~describe this method here~~
    }
  }
}

\examples{
## print | show
print(timeCalendar())
}

\keyword{chron}
