\name{myUnits}


\title{Frequency of date/time Units}


\alias{myUnits}


\description{

    A variable with the frequency of date/units.

}



\value{

    returns the date/time units, a character value.
    By default \code{"days"}.

}



\examples{
## myUnits -
   getRmetricsOptions("myUnits")
}


\keyword{chron}

