\name{DaylightSavingTime}
\alias{DaylightSavingTime}

\alias{Abidjan}
\alias{Accra}
\alias{Adak}
\alias{Addis_Ababa}
\alias{Adelaide}
\alias{Aden}
\alias{Algiers}
\alias{Almaty}
\alias{Amman}
\alias{Amsterdam}
\alias{Anadyr}
\alias{Anchorage}
\alias{Andorra}
\alias{Anguilla}
\alias{Antananarivo}
\alias{Antigua}
\alias{Apia}
\alias{Aqtau}
\alias{Aqtobe}
\alias{Araguaina}
\alias{Aruba}
\alias{Ashgabat}
\alias{Asmara}
\alias{Asuncion}
\alias{Athens}
\alias{Atikokan}
\alias{Auckland}
\alias{Azores}
\alias{Baghdad}
\alias{Bahia}
\alias{Bahrain}
\alias{Baku}
\alias{Bamako}
\alias{Bangkok}
\alias{Bangui}
\alias{Banjul}
\alias{Barbados}
\alias{Beirut}
\alias{Belem}
\alias{Belgrade}
\alias{Belize}
\alias{Berlin}
\alias{Bermuda}
\alias{Bishkek}
\alias{Bissau}
\alias{Blanc-Sablon}
\alias{Blantyre}
\alias{Boa_Vista}
\alias{Bogota}
\alias{Boise}
\alias{Bratislava}
\alias{Brazzaville}
\alias{Brisbane}
\alias{Broken_Hill}
\alias{Brunei}
\alias{Brussels}
\alias{Bucharest}
\alias{Budapest}
\alias{Buenos_Aires}
\alias{BuenosAires}
\alias{Bujumbura}
\alias{Cairo}
\alias{Calcutta}
\alias{Cambridge_Bay}
\alias{Campo_Grande}
\alias{Canary}
\alias{Cancun}
\alias{Cape_Verde}
\alias{Caracas}
\alias{Casablanca}
\alias{Casey}
\alias{Catamarca}
\alias{Cayenne}
\alias{Cayman}
\alias{Center}
\alias{Ceuta}
\alias{Chagos}
\alias{Chatham}
\alias{Chicago}
\alias{Chihuahua}
\alias{Chisinau}
\alias{Choibalsan}
\alias{Chongqing}
\alias{Christmas}
\alias{Cocos}
\alias{Colombo}
\alias{Comoro}
\alias{Conakry}
\alias{Copenhagen}
\alias{Cordoba}
\alias{Costa_Rica}
\alias{Cuiaba}
\alias{Curacao}
\alias{Currie}
\alias{Dakar}
\alias{Damascus}
\alias{Danmarkshavn}
\alias{Dar_es_Salaam}
\alias{Darwin}
\alias{Davis}
\alias{Dawson}
\alias{Dawson_Creek}
\alias{Denver}
\alias{Detroit}
\alias{Dhaka}
\alias{Dili}
\alias{Djibouti}
\alias{Dominica}
\alias{Douala}
\alias{Dubai}
\alias{Dublin}
\alias{DumontDUrville}
\alias{Dushanbe}
\alias{Eastern}
\alias{Edmonton}
\alias{Efate}
\alias{Eirunepe}
\alias{El_Aaiun}
\alias{El_Salvador}
\alias{Enderbury}
\alias{Eucla}
\alias{Fakaofo}
\alias{Faroe}
\alias{Fiji}
\alias{Fortaleza}
\alias{Frankfurt}
\alias{Freetown}
\alias{Funafuti}
\alias{Gaborone}
\alias{Galapagos}
\alias{Gambier}
\alias{Gaza}
\alias{Gibraltar}
\alias{Glace_Bay}
\alias{Godthab}
\alias{Goose_Bay}
\alias{Grand_Turk}
\alias{Grenada}
\alias{Guadalcanal}
\alias{Guadeloupe}
\alias{Guam}
\alias{Guatemala}
\alias{Guayaquil}
\alias{Guernsey}
\alias{Guyana}
\alias{Halifax}
\alias{Harare}
\alias{Harbin}
\alias{Havana}
\alias{Helsinki}
\alias{Hermosillo}
\alias{Hobart}
\alias{Hong_Kong}
\alias{HongKong}
\alias{Honolulu}
\alias{Hovd}
\alias{Indianapolis}
\alias{Inuvik}
\alias{Iqaluit}
\alias{Irkutsk}
\alias{Isle_of_Man}
\alias{Istanbul}
\alias{Jakarta}
\alias{Jamaica}
\alias{Jayapura}
\alias{Jersey}
\alias{Jerusalem}
\alias{Johannesburg}
\alias{Johnston}
\alias{Jujuy}
\alias{Juneau}
\alias{Kabul}
\alias{Kaliningrad}
\alias{Kamchatka}
\alias{Kampala}
\alias{Karachi}
\alias{Kashgar}
\alias{Katmandu}
\alias{Kerguelen}
\alias{Khartoum}
\alias{Kiev}
\alias{Kigali}
\alias{Kinshasa}
\alias{Kiritimati}
\alias{Knox}
\alias{Kosrae}
\alias{Krasnoyarsk}
\alias{Kuala_Lumpur}
\alias{KualaLumpur}
\alias{Kuching}
\alias{Kuwait}
\alias{Kwajalein}
\alias{La_Paz}
\alias{La_Rioja}
\alias{Lagos}
\alias{Libreville}
\alias{Lima}
\alias{Lindeman}
\alias{Lisbon}
\alias{Ljubljana}
\alias{Lome}
\alias{London}
\alias{Longyearbyen}
\alias{Lord_Howe}
\alias{Los_Angeles}
\alias{LosAngeles}
\alias{Louisville}
\alias{Luanda}
\alias{Lubumbashi}
\alias{Lusaka}
\alias{Luxembourg}
\alias{Macau}
\alias{Maceio}
\alias{Madeira}
\alias{Madrid}
\alias{Magadan}
\alias{Mahe}
\alias{Majuro}
\alias{Makassar}
\alias{Malabo}
\alias{Maldives}
\alias{Malta}
\alias{Managua}
\alias{Manaus}
\alias{Manila}
\alias{Maputo}
\alias{Marengo}
\alias{Mariehamn}
\alias{Marigot}
\alias{Marquesas}
\alias{Martinique}
\alias{Maseru}
\alias{Mauritius}
\alias{Mawson}
\alias{Mayotte}
\alias{Mazatlan}
\alias{Mbabane}
\alias{McMurdo}
\alias{Melbourne}
\alias{Mendoza}
\alias{Menominee}
\alias{Merida}
\alias{Mexico_City}
\alias{MexicoCity}
\alias{Midway}
\alias{Minsk}
\alias{Miquelon}
\alias{Mogadishu}
\alias{Monaco}
\alias{Moncton}
\alias{Monrovia}
\alias{Monterrey}
\alias{Montevideo}
\alias{Monticello}
\alias{Montreal}
\alias{Montserrat}
\alias{Moscow}
\alias{Muscat}
\alias{Nairobi}
\alias{Nassau}
\alias{Nauru}
\alias{Ndjamena}
\alias{New_Salem}
\alias{New_York}
\alias{NewYork}
\alias{Niamey}
\alias{Nicosia}
\alias{Nipigon}
\alias{Niue}
\alias{Nome}
\alias{Norfolk}
\alias{Noronha}
\alias{Nouakchott}
\alias{Noumea}
\alias{Novosibirsk}
\alias{Omsk}
\alias{Oral}
\alias{Oslo}
\alias{Ouagadougou}
\alias{Pacific}
\alias{Pago_Pago}
\alias{Palau}
\alias{Palmer}
\alias{Panama}
\alias{Pangnirtung}
\alias{Paramaribo}
\alias{Paris}
\alias{Perth}
\alias{Petersburg}
\alias{Phnom_Penh}
\alias{Phoenix}
\alias{Pitcairn}
\alias{Podgorica}
\alias{Ponape}
\alias{Pontianak}
\alias{Port_Moresby}
\alias{Port_of_Spain}
\alias{Port-au-Prince}
\alias{Porto_Velho}
\alias{Porto-Novo}
\alias{Prague}
\alias{Puerto_Rico}
\alias{Pyongyang}
\alias{Qatar}
\alias{Qyzylorda}
\alias{Rainy_River}
\alias{Rangoon}
\alias{Rankin_Inlet}
\alias{Rarotonga}
\alias{Recife}
\alias{Regina}
\alias{Resolute}
\alias{Reunion}
\alias{Reykjavik}
\alias{Riga}
\alias{Rio_Branco}
\alias{Rio_Gallegos}
\alias{Riyadh}
\alias{Rome}
\alias{Rothera}
\alias{Saigon}
\alias{Saipan}
\alias{Sakhalin}
\alias{Samara}
\alias{Samarkand}
\alias{San_Juan}
\alias{San_Marino}
\alias{Santiago}
\alias{Santo_Domingo}
\alias{Sao_Paulo}
\alias{Sao_Tome}
\alias{Sarajevo}
\alias{Scoresbysund}
\alias{Seoul}
\alias{Shanghai}
\alias{Shiprock}
\alias{Simferopol}
\alias{Singapore}
\alias{Skopje}
\alias{Sofia}
\alias{South_Georgia}
\alias{South_Pole}
\alias{St_Barthelemy}
\alias{St_Helena}
\alias{St_Johns}
\alias{St_Kitts}
\alias{St_Lucia}
\alias{St_Thomas}
\alias{St_Vincent}
\alias{Stanley}
\alias{Stockholm}
\alias{Swift_Current}
\alias{Sydney}
\alias{Syowa}
\alias{Tahiti}
\alias{Taipei}
\alias{Tallinn}
\alias{Tarawa}
\alias{Tashkent}
\alias{Tbilisi}
\alias{Tegucigalpa}
\alias{Tehran}
\alias{Tell_City}
\alias{Thimphu}
\alias{Thule}
\alias{Thunder_Bay}
\alias{Tijuana}
\alias{Tirane}
\alias{Tokyo}
\alias{Tongatapu}
\alias{Toronto}
\alias{Tortola}
\alias{Tripoli}
\alias{Truk}
\alias{Tucuman}
\alias{Tunis}
\alias{Ulaanbaatar}
\alias{Urumqi}
\alias{Ushuaia}
\alias{Uzhgorod}
\alias{Vaduz}
\alias{Vancouver}
\alias{Vatican}
\alias{Vevay}
\alias{Vienna}
\alias{Vientiane}
\alias{Vilnius}
\alias{Vincennes}
\alias{Vladivostok}
\alias{Volgograd}
\alias{Vostok}
\alias{Wake}
\alias{Wallis}
\alias{Warsaw}
\alias{Whitehorse}
\alias{Winamac}
\alias{Windhoek}
\alias{Winnipeg}
\alias{Yakutat}
\alias{Yakutsk}
\alias{Yekaterinburg}
\alias{Yellowknife}
\alias{Yerevan}
\alias{Zagreb}
\alias{Zaporozhye}
\alias{Zurich}

\alias{AST}
\alias{CET}
\alias{CST}
\alias{EET}
\alias{EST}
\alias{MST}
\alias{PST}


\title{Daylight Saving Time Rules}

\usage{

Abidjan()
Accra()
Adak()
Addis_Ababa()
Adelaide()
Aden()
Algiers()
Almaty()
Amman()
Amsterdam()
Anadyr()
Anchorage()
Andorra()
Anguilla()
Antananarivo()
Antigua()
Apia()
Aqtau()
Aqtobe()
Araguaina()
Aruba()
Ashgabat()
Asmara()
Asuncion()
Athens()
Atikokan()
Auckland()
Azores()
Baghdad()
Bahia()
Bahrain()
Baku()
Bamako()
Bangkok()
Bangui()
Banjul()
Barbados()
Beirut()
Belem()
Belgrade()
Belize()
Berlin()
Bermuda()
Bishkek()
Bissau()
`Blanc-Sablon`()
Blantyre()
Boa_Vista()
Bogota()
Boise()
Bratislava()
Brazzaville()
Brisbane()
Broken_Hill()
Brunei()
Brussels()
Bucharest()
Budapest()
Buenos_Aires()
BuenosAires()
Bujumbura()
Cairo()
Calcutta()
Cambridge_Bay()
Campo_Grande()
Canary()
Cancun()
Cape_Verde()
Caracas()
Casablanca()
Casey()
Catamarca()
Cayenne()
Cayman()
Center()
Ceuta()
Chagos()
Chatham()
Chicago()
Chihuahua()
Chisinau()
Choibalsan()
Chongqing()
Christmas()
Cocos()
Colombo()
Comoro()
Conakry()
Copenhagen()
Cordoba()
Costa_Rica()
Cuiaba()
Curacao()
Currie()
Dakar()
Damascus()
Danmarkshavn()
Dar_es_Salaam()
Darwin()
Davis()
Dawson()
Dawson_Creek()
Denver()
Detroit()
Dhaka()
Dili()
Djibouti()
Dominica()
Douala()
Dubai()
Dublin()
DumontDUrville()
Dushanbe()
Eastern()
Edmonton()
Efate()
Eirunepe()
El_Aaiun()
El_Salvador()
Enderbury()
Eucla()
Fakaofo()
Faroe()
Fiji()
Fortaleza()
Frankfurt()
Freetown()
Funafuti()
Gaborone()
Galapagos()
Gambier()
Gaza()
Gibraltar()
Glace_Bay()
Godthab()
Goose_Bay()
Grand_Turk()
Grenada()
Guadalcanal()
Guadeloupe()
Guam()
Guatemala()
Guayaquil()
Guernsey()
Guyana()
Halifax()
Harare()
Harbin()
Havana()
Helsinki()
Hermosillo()
Hobart()
Hong_Kong()
HongKong()
Honolulu()
Hovd()
Indianapolis()
Inuvik()
Iqaluit()
Irkutsk()
Isle_of_Man()
Istanbul()
Jakarta()
Jamaica()
Jayapura()
Jersey()
Jerusalem()
Johannesburg()
Johnston()
Jujuy()
Juneau()
Kabul()
Kaliningrad()
Kamchatka()
Kampala()
Karachi()
Kashgar()
Katmandu()
Kerguelen()
Khartoum()
Kiev()
Kigali()
Kinshasa()
Kiritimati()
Knox()
Kosrae()
Krasnoyarsk()
Kuala_Lumpur()
KualaLumpur()
Kuching()
Kuwait()
Kwajalein()
La_Paz()
La_Rioja()
Lagos()
Libreville()
Lima()
Lindeman()
Lisbon()
Ljubljana()
Lome()
London()
Longyearbyen()
Lord_Howe()
Los_Angeles()
LosAngeles()
Louisville()
Luanda()
Lubumbashi()
Lusaka()
Luxembourg()
Macau()
Maceio()
Madeira()
Madrid()
Magadan()
Mahe()
Majuro()
Makassar()
Malabo()
Maldives()
Malta()
Managua()
Manaus()
Manila()
Maputo()
Marengo()
Mariehamn()
Marigot()
Marquesas()
Martinique()
Maseru()
Mauritius()
Mawson()
Mayotte()
Mazatlan()
Mbabane()
McMurdo()
Melbourne()
Mendoza()
Menominee()
Merida()
Mexico_City()
MexicoCity()
Midway()
Minsk()
Miquelon()
Mogadishu()
Monaco()
Moncton()
Monrovia()
Monterrey()
Montevideo()
Monticello()
Montreal()
Montserrat()
Moscow()
Muscat()
Nairobi()
Nassau()
Nauru()
Ndjamena()
New_Salem()
New_York()
NewYork()
Niamey()
Nicosia()
Nipigon()
Niue()
Nome()
Norfolk()
Noronha()
Nouakchott()
Noumea()
Novosibirsk()
Omsk()
Oral()
Oslo()
Ouagadougou()
Pacific()
Pago_Pago()
Palau()
Palmer()
Panama()
Pangnirtung()
Paramaribo()
Paris()
Perth()
Petersburg()
Phnom_Penh()
Phoenix()
Pitcairn()
Podgorica()
Ponape()
Pontianak()
Port_Moresby()
Port_of_Spain()
`Port-au-Prince`()
Porto_Velho()
`Porto-Novo`()
Prague()
Puerto_Rico()
Pyongyang()
Qatar()
Qyzylorda()
Rainy_River()
Rangoon()
Rankin_Inlet()
Rarotonga()
Recife()
Regina()
Resolute()
Reunion()
Reykjavik()
Riga()
Rio_Branco()
Rio_Gallegos()
Riyadh()
Rome()
Rothera()
Saigon()
Saipan()
Sakhalin()
Samara()
Samarkand()
San_Juan()
San_Marino()
Santiago()
Santo_Domingo()
Sao_Paulo()
Sao_Tome()
Sarajevo()
Scoresbysund()
Seoul()
Shanghai()
Shiprock()
Simferopol()
Singapore()
Skopje()
Sofia()
South_Georgia()
South_Pole()
St_Barthelemy()
St_Helena()
St_Johns()
St_Kitts()
St_Lucia()
St_Thomas()
St_Vincent()
Stanley()
Stockholm()
Swift_Current()
Sydney()
Syowa()
Tahiti()
Taipei()
Tallinn()
Tarawa()
Tashkent()
Tbilisi()
Tegucigalpa()
Tehran()
Tell_City()
Thimphu()
Thule()
Thunder_Bay()
Tijuana()
Tirane()
Tokyo()
Tongatapu()
Toronto()
Tortola()
Tripoli()
Truk()
Tucuman()
Tunis()
Ulaanbaatar()
Urumqi()
Ushuaia()
Uzhgorod()
Vaduz()
Vancouver()
Vatican()
Vevay()
Vienna()
Vientiane()
Vilnius()
Vincennes()
Vladivostok()
Volgograd()
Vostok()
Wake()
Wallis()
Warsaw()
Whitehorse()
Winamac()
Windhoek()
Winnipeg()
Yakutat()
Yakutsk()
Yekaterinburg()
Yellowknife()
Yerevan()
Zagreb()
Zaporozhye()
Zurich()

AST()
CET()
CST()
EET()
EST()
MST()
PST()

}

\description{

  Functions for about 400 cities and regions which return daylight
  saving time rules and time zone offsets.
  
}

%\details{
%
%}

\note{

  There are currently two synonyms available "Pacific" for Los Angeles
  and "Eastern" for New York. 

  Specific time zones (\code{AST}, \code{CET}, \code{CST}, \code{EET},
  \code{EST}, \code{MST} and \code{PST}) are also available.
  
  Note we leave the space in all double named cities like New York 
  or Hong Kong and use an underscore for it.
  
  All the entries are retrieved from the tzdata library which is 
  available under GNU GPL licence. 

}

\examples{
## DST rules for Zurich
head(Zurich())
tail(Zurich())

## list all available centers
listFinCenter()
}

\keyword{data}
