\name{show-methods}
\alias{show-methods}

\docType{methods}

\alias{show,ANY-method}
\alias{show,timeDate-method}
\alias{show.timeDate}


\title{Show methods}

\description{
  Show methods for \code{"timeDate"} objects. 
}

\section{Methods}{
  \describe{

    \item{object = "ANY"}{
      Generic function.
    }
    \item{object = "timeDate"}{
      Print function for objects of class \code{"timeDate"}.
    } 
  }
}

\examples{
## print | show
print(timeCalendar())
}

\keyword{chron}
